# lib/api-zod/src/generated/types/aiGenerateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MatricBlog API
 * OpenAPI spec version: 0.1.0
 */
import type { AiGenerateResultData } from "./aiGenerateResultData";

export interface AiGenerateResult {
  success: boolean;
  data: AiGenerateResultData;
}