# lib/api-zod/src/generated/types/aiImproveInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MatricBlog API
 * OpenAPI spec version: 0.1.0
 */

export interface AiImproveInput {
  html: string;
  focusKeyword?: string;
  instruction?: string;
}