# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MatricBlog API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryInput {
  name: string;
  /** @nullable */
  name_ar?: string | null;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
}