# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MatricBlog API
 * OpenAPI spec version: 0.1.0
 */
import type { PostInputStatus } from "./postInputStatus";

export interface PostInput {
  title: string;
  /** @nullable */
  title_ar?: string | null;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  content_ar?: string | null;
  /** @nullable */
  featured_image?: string | null;
  /** @nullable */
  category_id?: string | null;
  status: PostInputStatus;
  /** @nullable */
  published_at?: string | null;
  /** @nullable */
  scheduled_at?: string | null;
  /** @nullable */
  meta_title?: string | null;
  /** @nullable */
  meta_description?: string | null;
  /** @nullable */
  meta_keywords?: string[] | null;
  /** @nullable */
  og_image?: string | null;
  /** @nullable */
  canonical_url?: string | null;
  /** @nullable */
  robots_meta?: string | null;
  /** @nullable */
  schema_type?: string | null;
  /** @nullable */
  reading_time?: number | null;
}